+++ src/Resources/contao/config/config.php
@@ -155,6 +155,8 @@ $GLOBALS['TL_CONFIG']['fernschach_hinweis_kontoauszug'] = 'Kein BdF-Mitglied';
 $GLOBALS['TL_CONFIG']['fernschach_resetUpdate_time'] = 86400;
 $GLOBALS['TL_CONFIG']['fernschach_membershipUpdate_time'] = 86400;
 $GLOBALS['TL_CONFIG']['fernschach_maintenanceUpdate_time'] = 43200;
+$GLOBALS['TL_CONFIG']['fernschach_intervall_memberbridgeCheck'] = 86400;
+$GLOBALS['TL_CONFIG']['fernschach_intervall_membershipsCheck'] = 86400;
 
 /**
  * -------------------------------------------------------------------------