+++ src/Resources/contao/languages/de/tl_fernschach_spieler.php
@@ -5,6 +5,7 @@
  */
 $GLOBALS['TL_LANG']['tl_fernschach_spieler']['id'] = array('ID', 'ID dieses Spielers');
 $GLOBALS['TL_LANG']['tl_fernschach_spieler']['tstamp'] = array('Letzte Änderung', 'Zeitstempel der letzten Änderung');
+$GLOBALS['TL_LANG']['tl_fernschach_spieler']['memberbridgeTime'] = array('Letzte Wartung', 'Zeitstempel der letzten Wartung');
 
 $GLOBALS['TL_LANG']['tl_fernschach_spieler']['archived_legend'] = 'Archivierung';
 $GLOBALS['TL_LANG']['tl_fernschach_spieler']['archived'] = array('Archiviert', 'Der Spieler wird in der Übersicht in grauer Schrift angezeigt.');